feat(dm): Setup sh collection
The sh collection provides the base shell dotfile configurations for a
minimum shell environment that follows the POSIX standard like the ASH
or DASH shell.

I like to use the polyglot PS1 project that provides a unified PS1
prompt for a whole lot of shell environments. It is very cool, have a
look for yourself <https://github.com/agkozak/polyglot>.

Signed-off-by: Jason Pena <jasonpena@awkless.com>