fix(dm): Alter method of obtaining submodules
The old way would not update to the newest changes found in any given
submodule, and would occasionally break the submodule itself when you
tried to manually update the submodule itself since it would revert back
to the previous changes for reasons I do not fully understand. Luckily,
this code fixes that (hopefully).

Signed-off-by: Jason Pena <jasonpena@awkless.com>